requirements.txt
web3==6.11.0
psycopg2-binary==2.9.9
requests==2.31.0
flask==3.0.0
flask-cors==4.0.0
discord-webhook==1.3.0
aiohttp==3.9.0
websockets==12.0
python-dotenv==1.0.0
gunicorn==21.2.0